Split out slow tests with tag `slow tests`
